one less go routine
